Check for API key before adding or removing entry

+++ app/api/marquee/route.ts
@@ -3,27 +3,37 @@ import {
   newMarqueeEntry,
   updateMarqueeEntry,
   deleteMarqueeEntries
-} from '@/functions/db/marquee'
+} from "@/functions/db/marquee"
 
 export async function POST(request: Request) {
+  const headers = request.headers
+  const authorization = headers.get("authorization")
+  const key = authorization?.split(" ")[1]
+  if (key !== process.env.MARQUEE_API_KEY)
+    return Response.json({ error: "Unauthorized" }, { status: 401 })
+
   const body = await request.json()
   if (!body.id || !body.nickname || !body.project)
-    return Response.json({ error: 'Missing required fields' }, { status: 400 })
+    return Response.json({ error: "Missing required fields" }, { status: 400 })
 
   // Not idiomatic HTTP, but we'll live with it for now
   const marquee = await getMarqueeEntry(body.id)
-  if (marquee)
-    await updateMarqueeEntry(body)
-  else
-    await newMarqueeEntry(body)
+  if (marquee) await updateMarqueeEntry(body)
+  else await newMarqueeEntry(body)
 
   return Response.json({ body })
 }
 
 export async function DELETE(request: Request) {
+  const headers = request.headers
+  const authorization = headers.get("authorization")
+  const key = authorization?.split(" ")[1]
+  if (key !== process.env.MARQUEE_API_KEY)
+    return Response.json({ error: "Unauthorized" }, { status: 401 })
+
   const body = await request.json()
   if (!body.id)
-    return Response.json({ error: 'Missing required fields' }, { status: 400 })
+    return Response.json({ error: "Missing required fields" }, { status: 400 })
 
   const numDeleted = await deleteMarqueeEntries(body.id)
   return Response.json({ count: numDeleted })
